pyproject.toml
[tool.poetry]
name = "pidioul"
version = "0.1.0"
description = ""
authors = ["Evann DREUMONT <53308142+LeGmask@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jupyter = "^1.0.0"
chess = "^1.9.3"
"discord.py" = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
